server/requirements.txt
fastapi
uvicorn
python-dotenv
authlib
httpx
starlette-session
motor
python-jose
cryptography
pytest
pytest-asyncio
pytest-cov
pytest-mock
